+++ src/main/java/ru/akvine/istochnik/services/converters/integer/NegativeIntegerConverter.java
@@ -6,9 +6,9 @@ import ru.akvine.istochnik.enums.ConverterType;
 import java.util.List;
 
 @Service
-public class NegativeIntegerConverter extends IntegerConverter<Integer, Double> {
+public class NegativeIntegerConverter extends IntegerConverter<Long, Double> {
     @Override
-    public List<Integer> convert(List<Integer> input, Double[] arguments) {
+    public List<Long> convert(List<Long> input, Double[] arguments) {
         return input.stream().map(value -> value * (-1)).toList();
     }
